requirements.txt
groq
Pillow
openai
faster-whisper
SpeechRecognition
google-generativeai
pyperclip
opencv-python
pyaudio
